// src/switch-slider/types.ts
import { SIZES } from './constants'

type Values<T> = T[keyof T]
type SwitchSize = Values<typeof SIZES>

export type SwitchSliderProps = {
    /**
     * Начальное состояние переключателя.
     *
     * Определяет, какое состояние будет установлено при первой загрузке:
     * включенное, выключенное или нейтральное.
     */
    initialState?: 'enabled' | 'disabled' | 'neutral'
    /**
     * Уникальный идентификатор для переключателя.
     *
     * Используется для связи с другими элементами и улучшения доступности.
     */
    id?: string
    /**
     * Имя переключателя для идентификации в группе элементов.
     *
     * Полезно для группировки элементов и отправки данных формы.
     */
    name?: string
    /**
     * Размер переключателя.
     *
     * Задает размер переключателя, выбираемый из предопределенных значений в `SIZES`.
     */
    size?: SwitchSize
    /**
     * Текст, сопровождающий переключатель.
     *
     * Используется для обозначения назначения переключателя.
     */
    label?: string
    /**
     * Состояние переключателя.
     *
     * Указывает, выбран ли переключатель в данный момент.
     */
    checked?: boolean
    /**
     * Отключение переключателя.
     *
     * Делает переключатель неактивным и недоступным для взаимодействия.
     */
    disabled?: boolean
    /**
     * Функция-обработчик, вызываемая при изменении состояния переключателя.
     *
     * Срабатывает при смене состояния переключателя, передавая событие
     * и данные о текущем состоянии.
     */
    onChange?: (
        event: React.ChangeEvent<HTMLInputElement>,
        data: { value: string; checked: boolean },
    ) => void
    /**
     * Кастомный CSS-класс для стилизации.
     *
     * Позволяет применить дополнительные стили к переключателю.
     */
    className?: string
    /**
     * Значение переключателя, используемое при отправке формы.
     *
     * Определяет значение, передаваемое при выборе переключателя.
     */
    value?: string
}